refactor: change names for some files and variables
