+++ index.ts
@@ -1,5 +1,5 @@
 import dotenv from "dotenv";
-import startApolloServer from "./src/server";
+import startApolloServer from "./src/app";
 
 dotenv.config();
 startApolloServer();
